+++ 1-settings/classColors.lua
@@ -1,0 +1,18 @@
+----------------------------------------------------------------------------------------------------------
+local A, L, V, P, G = unpack(select(2, ...)); -- Import: Engine, Locales, PrivateDB, ProfileDB, GlobalDB
+----------------------------------------------------------------------------------------------------------
+--[[
+    !! Used by CRIndicator
+]]
+Arch_classColors = {
+    ["DeathKnight"] = '|cffC41F3B',    
+    ["Druid"] = '|cffFF7D0A',
+    ["Hunter"] = '|cffA9D271',
+    ["Mage"] = '|cff40C7EB',
+    ["Paladin"] = '|cffF58CBA',
+    ["Priest"] = '|cffFFFFFF',
+    ["Rogue"] = '|cffFFF569',
+    ["Shaman"] = '|cff0070DE',
+    ["Warlock"] = '|cff8787ED',
+    ["Warrior"] = '|cffC79C6E',
+}