+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # ChangeLog 
 `Started 29-Jul-2020`
 
+[0.9.6]
+Rework CRIndicator, set database for spells instead of data function
+Rework CRIndicator, set class colors for list
+Added CRIndicator, remove people who are not in raid anymore
+Update CRIndicator, using spell names instead of id
+Added CRIndicator, removing bears by tracking mangle spell -- disabled
+[0.9.5]
+Added PlayerDB whitelist, blacklist
+Fixed whitelist, blacklist comma bug
+Fixed PlayerDB esc disabling
+Added PlayerDB offline player entry with prefix `/`
 [0.9.4] 
 Added tactics adjustability, raid warnings, silent reminders